Add files via upload
This is the first ever game I have made to practice implementing AI and making video games. This game is a small duel which includes an AI bot driven by a heuristic alpha-beta search algorithm. This game was coded in python through Jupyter Lab and I used the pygame library to implement animations and gameplay.